+++ docs/conf.py
@@ -53,7 +53,6 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.doctest',
     'sphinx.ext.intersphinx',
-    'sphinx.ext.todo',
     'sphinx.ext.viewcode',
     'sphinxcontrib.autoprogram',
     # enable pylons_sphinx_latesturl when this branch is no longer "latest"
@@ -124,9 +123,6 @@ exclude_patterns = ['_themes/README.rst', ]
 # unit titles (such as .. function::).
 add_module_names = False
 
-# Add support for todo items
-todo_include_todos = True
-
 # The name of the Pygments (syntax highlighting) style to use.
 #pygments_style = book and 'bw' or 'tango'
 if book:
@@ -175,8 +171,6 @@ html_last_updated_fmt = '%b %d, %Y'
 
 # Do not use smart quotes.
 smartquotes = False
-# Remove next line when RTD goes to Sphinx==1.6.6
-html_use_smartypants = False
 
 # Output file base name for HTML help builder.
 htmlhelp_basename = 'pyramid'
@@ -190,7 +184,10 @@ latex_paper_size = 'letter'
 # The font size ('10pt', '11pt' or '12pt').
 latex_font_size = '10pt'
 
-latex_additional_files = ['_static/latex-note.png', '_static/latex-warning.png']
+latex_additional_files = [
+    '_static/latex-note.png',
+    '_static/latex-warning.png',
+]
 
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, document class [howto/manual]).
@@ -220,6 +217,17 @@ latex_domain_indices = False
 _PREAMBLE = r"""
 \usepackage[]{geometry}
 \geometry{bindingoffset=0.45in,textheight=7.25in,hdivide={0.5in,*,0.75in},vdivide={1in,7.25in,1in},papersize={7.5in,9.25in}}
+
+%XeLaTeX packages
+\usepackage{xltxtra}
+\usepackage{fontspec} %Font package
+\usepackage{xunicode}
+
+%Select fonts
+\setmainfont[Mapping=tex-text]{nimbusserif}
+\setsansfont[Mapping=tex-text]{nimbussans}
+\setmonofont{nimbusmono}
+
 \hypersetup{
     colorlinks=true,
     linkcolor=black,
@@ -276,28 +284,22 @@ _PREAMBLE = r"""
 \definecolor{VerbatimColor}{rgb}{1,1,1}
 \definecolor{VerbatimBorderColor}{rgb}{1,1,1}
 
-\makeatletter
-\renewcommand{\py@noticestart@warning}{\py@heavybox}
-\renewcommand{\py@noticeend@warning}{\py@endheavybox}
-\renewcommand{\py@noticestart@note}{\py@heavybox}
-\renewcommand{\py@noticeend@note}{\py@endheavybox}
-\makeatother
-
 % icons in note and warning boxes
 \usepackage{ifthen}
-% Keep a copy of the original notice environment
-\let\origbeginnotice\notice
-\let\origendnotice\endnotice
 
-% Redefine the notice environment so we can add our own code to it
-\renewenvironment{notice}[2]{%
-  \origbeginnotice{#1}{}% equivalent to original \begin{notice}{#1}{#2}
+% Keep a copy of the original sphinxadmonition environment
+\let\origbeginadmon\sphinxadmonition
+\let\origendadmon\endsphinxadmonition
+
+% Redefine the sphinxadmonition environment so we can add our own code to it
+\renewenvironment{sphinxadmonition}[2]{%
+  \origbeginadmon{#1}{}% equivalent to original \begin{sphinxadmonition}{#1}{#2}
   % load graphics
   \ifthenelse{\equal{#1}{warning}}{\includegraphics{latex-warning.png}}{}
   \ifthenelse{\equal{#1}{note}}{\includegraphics{latex-note.png}}{}
   % etc.
-}{%
-  \origendnotice% equivalent to original \end{notice}
+  }{%
+\origendadmon % equivalent to original \end{sphinxadmonition}
 }
 
 % try to prevent code-block boxes from splitting across pages
@@ -318,7 +320,6 @@ _PREAMBLE = r"""
 
 latex_elements = {
     'preamble': _PREAMBLE,
-    'wrapperclass': 'book',
     'date': '',
     'releasename': 'Version',
     'title': r'The Pyramid Web Framework',
@@ -342,17 +343,6 @@ def frontmatter(name, arguments, options, content, lineno,
                 content_offset, block_text, state, state_machine):
     return [nodes.raw(
         '',
-        r"""
-\frontmatter
-% prevent part/chapter/section numbering
-\setcounter{secnumdepth}{-2}
-% suppress headers
-\pagestyle{plain}
-% reset page counter
-\setcounter{page}{1}
-% suppress first toc pagenum
-\addtocontents{toc}{\protect\thispagestyle{empty}}
-""",
         format='latex')]
 
 
@@ -360,26 +350,14 @@ def mainmatter(name, arguments, options, content, lineno,
                content_offset, block_text, state, state_machine):
     return [nodes.raw(
         '',
-        r"""
-\mainmatter
-% allow part/chapter/section numbering
-\setcounter{secnumdepth}{2}
-% get headers back
-\pagestyle{fancy}
-\fancyhf{}
-\renewcommand{\headrulewidth}{0.5pt}
-\renewcommand{\footrulewidth}{0pt}
-\fancyfoot[C]{\thepage}
-\fancyhead[RO]{\rightmark}
-\fancyhead[LE]{\leftmark}
-""",
         format='latex')]
 
 
 def backmatter(name, arguments, options, content, lineno,
               content_offset, block_text, state, state_machine):
-    return [nodes.raw('', '\\backmatter\n\\setcounter{secnumdepth}{-1}\n',
-                      format='latex')]
+    return [nodes.raw(
+        '',
+        format='latex')]
 
 
 def app_role(role, rawtext, text, lineno, inliner, options={}, content=[]):