+++ pyramid/httpexceptions.py
@@ -7,7 +7,7 @@ single HTTP status code.  Each class is a subclass of the
 :class:`~HTTPException`.  Each exception class is also a :term:`response`
 object.
 
-Each exception class has a status code according to :rfc:`2068`:
+Each exception class has a status code according to :rfc:`2068` or :rfc:`7538`:
 codes with 100-300 are not really errors; 400s are client errors,
 and 500s are server errors.
 
@@ -29,6 +29,7 @@ Exception
       * 304 - HTTPNotModified
       * 305 - HTTPUseProxy
       * 307 - HTTPTemporaryRedirect
+      * 308 - HTTPPermanentRedirect
     HTTPError
       HTTPClientError
         * 400 - HTTPBadRequest
@@ -121,10 +122,11 @@ passed to the exception's constructor.
 
 The subclasses of :class:`~_HTTPMove`
 (:class:`~HTTPMultipleChoices`, :class:`~HTTPMovedPermanently`,
-:class:`~HTTPFound`, :class:`~HTTPSeeOther`, :class:`~HTTPUseProxy` and
-:class:`~HTTPTemporaryRedirect`) are redirections that require a ``Location``
-field. Reflecting this, these subclasses have one additional keyword argument:
-``location``, which indicates the location to which to redirect.
+:class:`~HTTPFound`, :class:`~HTTPSeeOther`, :class:`~HTTPUseProxy`,
+:class:`~HTTPTemporaryRedirect`, and :class: `~HTTPPermanentRedirect) are
+redirections that require a ``Location`` field. Reflecting this, these
+subclasses have one additional keyword argument: ``location``,
+which indicates the location to which to redirect.
 """
 import json
 
@@ -133,7 +135,7 @@ from string import Template
 from zope.interface import implementer
 
 from webob import html_escape as _html_escape
-from webob.acceptparse import MIMEAccept
+from webob.acceptparse import create_accept_header
 
 from pyramid.compat import (
     class_types,
@@ -250,10 +252,12 @@ ${body}''')
             html_comment = ''
             comment = self.comment or ''
             accept_value = environ.get('HTTP_ACCEPT', '')
-            accept = MIMEAccept(accept_value)
+            accept = create_accept_header(accept_value)
             # Attempt to match text/html or application/json, if those don't
             # match, we will fall through to defaulting to text/plain
-            match = accept.best_match(['text/html', 'application/json'])
+            acceptable = accept.acceptable_offers(['text/html', 'application/json'])
+            acceptable = [offer[0] for offer in acceptable] + ['text/plain']
+            match = acceptable[0]
 
             if match == 'text/html':
                 self.content_type = 'text/html'
@@ -593,6 +597,19 @@ class HTTPTemporaryRedirect(_HTTPMove):
     code = 307
     title = 'Temporary Redirect'
 
+class HTTPPermanentRedirect(_HTTPMove):
+    """
+    subclass of :class:`~_HTTPMove`
+
+    This indicates that the requested resource resides permanently
+    under a different URI and that the request method must not be
+    changed.
+
+    code: 308, title: Permanent Redirect
+    """
+    code = 308
+    title = 'Permanent Redirect'
+
 ############################################################
 ## 4xx client error
 ############################################################
@@ -606,7 +623,8 @@ class HTTPClientError(HTTPError):
     a bug.  A server-side traceback is not warranted.  Unless specialized,
     this is a '400 Bad Request'
     """
-    pass
+    code = 400
+    title = 'Bad Request'
 
 class HTTPBadRequest(HTTPClientError):
     """
@@ -617,8 +635,6 @@ class HTTPBadRequest(HTTPClientError):
 
     code: 400, title: Bad Request
     """
-    code = 400
-    title = 'Bad Request'
     explanation = ('The server could not comply with the request since '
                    'it is either malformed or otherwise incorrect.')
 
@@ -1032,11 +1048,18 @@ class HTTPServerError(HTTPError):
     This is an error condition in which the server is presumed to be
     in-error.  Unless specialized, this is a '500 Internal Server Error'.
     """
-    pass
-
-class HTTPInternalServerError(HTTPServerError):
     code = 500
     title = 'Internal Server Error'
+
+class HTTPInternalServerError(HTTPServerError):
+    """
+    subclass of :class:`~HTTPServerError`
+
+    This indicates that the server encountered an unexpected condition
+    which prevented it from fulfilling the request.
+
+    code: 500, title: Internal Server Error
+    """
     explanation = (
         'The server has either erred or is incapable of performing '
         'the requested operation.')
@@ -1150,6 +1173,7 @@ for name, value in list(globals().items()):
     if (
             isinstance(value, class_types) and
             issubclass(value, HTTPException) and
+            value not in {HTTPClientError, HTTPServerError} and
             not name.startswith('_')
     ):
         code = getattr(value, 'code', None)